Rename file about-nav.vue to AboutNav.vue
